chore: move SC tests under contracts folder. Fix CODEOWNERS
