+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/contract/hip906/HbarAllowanceApprovalTest.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.hedera.services.bdd.suites.hip906;
+package com.hedera.services.bdd.suites.contract.hip906;
 
 import static com.hedera.services.bdd.junit.TestTags.SMART_CONTRACT;
 import static com.hedera.services.bdd.spec.HapiPropertySource.asEntityString;
