+++ hedera-node/test-clients/src/main/java/com/hedera/services/bdd/suites/contract/records/ContractRecordsSanityCheckSuite.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.hedera.services.bdd.suites.records;
+package com.hedera.services.bdd.suites.contract.records;
 
 import static com.hedera.services.bdd.junit.ContextRequirement.SYSTEM_ACCOUNT_BALANCES;
 import static com.hedera.services.bdd.junit.TestTags.SMART_CONTRACT;
